refactor: 从 command.py 拆分 admin dispatcher 到独立模块
- 新增 persona/admin.py 容纳 16 个 admin 子命令处理器，
  独立 AdminDispatcher 类注入 bot/app/data_store 依赖
- command.py 移除 _admin_code/_admin_whitelist/_admin_debug 等
  16 个 admin 方法及 _format_relationship_base/_get_relationship_for_display
  辅助方法，瘦身约 460 行
- _admin_debug 通过 kwargs 传入 command 实时状态
  (tick_pending/daily_pending/observation_status)，保持跨模块边界清晰
